Add thefuck, error handling for swap

+++ apps/swap.sh
@@ -6,6 +6,23 @@ if [[ -a /tmp/ee_swap ]] ; then
   exit 1
 fi
 
+# if $1 is neither a file nor a directory
+if [ ! -f $1 ] && [ ! -d $1 ] ; then
+  echo "swap: $1: No such file or directory"
+  file_not_found=1
+fi
+
+# if $2 is neither a file nor a directory
+if [ ! -f $2 ] && [ ! -d $2 ]; then
+  echo "swap: $2: No such file or directory"
+  file_not_found=1
+fi
+
+if [[ $file_not_found ]] ; then
+  unset file_not_found
+  exit 2
+fi
+
 # teacup
 cp $1 /tmp/ee_swap
 rm -rf $1
